[chore/util] 비슷한 성격의 Util 파일 통합
