+++ app/src/main/java/com/mukmuk/todori/util/DateUtil.kt
@@ -3,6 +3,9 @@ package com.mukmuk.todori.util
 import android.os.Build
 import androidx.annotation.RequiresApi
 import kotlinx.datetime.DayOfWeek
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import java.time.temporal.ChronoUnit
 import kotlinx.datetime.LocalDate as KxLocalDate
 import java.time.LocalDate as JtLocalDate
 
@@ -36,3 +39,15 @@ fun DayOfWeek.toKoreanString(): String {
         DayOfWeek.SUNDAY -> "일"
     }
 }
+@RequiresApi(Build.VERSION_CODES.O)
+fun calculateDDayValue(dueDateString: String): Int? {
+    return try {
+        val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+        val dueDate = LocalDate.parse(dueDateString, formatter)
+        val today = LocalDate.now()
+        ChronoUnit.DAYS.between(today, dueDate).toInt()
+    } catch (e: Exception) {
+        null
+    }
+}
+
